Check detailed resource authorization for actions

+++ web/clui/routes/org.go
@@ -53,7 +53,6 @@ func (a *OrgAdmin) Create(ctx context.Context, name, owner string) (org *model.O
 		log.Println("DB failed to create organization ", err)
 		return
 	}
-	log.Println("$$$$$$$$$$$$$$$$$$$$$$$$$$$$ org = ", org.Name, org.ID)
 	secGroup, err := secgroupAdmin.Create(ctx, name, true, org.ID)
 	if err != nil {
 		log.Println("Failed to create security group", err)
@@ -75,6 +74,13 @@ func (a *OrgAdmin) Create(ctx context.Context, name, owner string) (org *model.O
 	if err != nil {
 		log.Println("DB failed to update orgabization default security group", err)
 		return
+	}
+	subnet, err := subnetAdmin.Create(ctx, name, "", "192.168.127.0", "255.255.255.0", "", "", "", "", "", org.ID)
+	if err != nil {
+		log.Println("Failed to create demo subnet", err)
+	} else {
+		subIDs := []int64{subnet.ID}
+		_, err = gatewayAdmin.Create(ctx, name, 0, 0, subIDs, org.ID)
 	}
 	return
 }
@@ -175,10 +181,34 @@ func (a *OrgAdmin) Delete(id int64) (err error) {
 		}
 	}()
 
+	count := 0
+	err = db.Model(&model.Interface{}).Where("owner = ?", id).Count(&count).Error
+	if err != nil {
+		log.Println("DB failed to query interfaces, %v", err)
+		return
+	}
+	if count > 0 {
+		log.Println("There are resources in this org", err)
+		err = fmt.Errorf("There are resources in this org")
+		return
+	}
 	err = db.Delete(&model.Member{}, `org_id = ?`, id).Error
 	if err != nil {
 		log.Println("DB failed to delete member, %v", err)
 		return
+	}
+	secgroups := []*model.SecurityGroup{}
+	err = db.Where("owner = ?", id).Find(&secgroups).Error
+	if err != nil {
+		log.Println("DB failed to query security groups", err)
+		return
+	}
+	for _, sg := range secgroups {
+		err = secgroupAdmin.Delete(sg.ID)
+		if err != nil {
+			log.Println("Can not delete security group", err)
+			return
+		}
 	}
 	err = db.Delete(&model.Organization{Model: model.Model{ID: id}}).Error
 	if err != nil {
